+++ src/main/scala/nl/knaw/dans/easy/stage/EasyStageDataset.scala
@@ -3,9 +3,12 @@ package nl.knaw.dans.easy.stage
 import java.io.File
 import java.net.URL
 
-import nl.knaw.dans.easy.stage.Constants._
-import nl.knaw.dans.easy.stage.FOXML._
+import nl.knaw.dans.easy.stage.lib._
+import Constants._
+import nl.knaw.dans.easy.stage.lib.{EasyItemContainerMd, EasyFileMetadata, JSON, FOXML}
+import FOXML._
 import nl.knaw.dans.easy.stage.Util._
+import nl.knaw.dans.easy.stage.lib.Util._
 import org.apache.commons.configuration.PropertiesConfiguration
 import org.apache.commons.io.FileUtils
 import org.slf4j.LoggerFactory
