+++ src/main/scala/nl/knaw/dans/easy/stage/Settings.scala
@@ -3,6 +3,8 @@ package nl.knaw.dans.easy.stage
 import java.io.File
 import java.net.URL
 
+import nl.knaw.dans.easy.stage.lib.Fedora
+
 case class Settings(ownerId: String,
                     submissionTimestamp: String,
                     bagStorageLocation: String,
