+++ src/main/scala/nl/knaw/dans/easy/stage/fileitem/FileItemConf.scala
@@ -1,8 +1,8 @@
-package nl.knaw.dans.easy.stage
+package nl.knaw.dans.easy.stage.fileitem
 
 import java.io.File
 import java.util.UUID
-
+import nl.knaw.dans.easy.stage.lib.Version
 import org.joda.time.DateTime
 import org.rogach.scallop.{ScallopConf, ValueConverter, singleArgConverter}
 import org.slf4j.LoggerFactory
