+++ src/main/scala/nl/knaw/dans/easy/stage/fileitem/FileItemCsv.scala
@@ -1,4 +1,4 @@
-package nl.knaw.dans.easy.stage
+package nl.knaw.dans.easy.stage.fileitem
 
 import java.io.InputStream
